+++ lib/src/hop_tasks/unit_test.dart
@@ -1,8 +1,11 @@
+@deprecated
 library hop_tasks.unit_test;
 
 import 'package:hop/hop_core.dart';
 import 'package:hop_unittest/hop_unittest.dart' as hut;
 
+/// **DEPRECATED**. Use `hop_unittest` package instead.
+///
 /// Creates a [Task] which runs the unit tests defined by [unitTestAction].
 ///
 /// [unitTestAction] should be in the form `void function()`.